# Optimize memtable(InlineSkipList) reverse scan