Signed-off-by: Little-Wallace liuwei@pingcap.com

When the iterator read keys in reverse order, each Prev() function cost O(log n) times. So I add prev pointer for every node in skiplist to improve the Prev() function.
Here is bench result.
```
time ./db_bench --db=./data --write_buffer_size=128000000 --num=320000 --benchmarks=fillseq,seekrandom --seek_nexts=100 --reads=1000000 --reverse_iterator=true
```
**master**:
```
DB path: [./data]
fillseq      :      14.732 micros/op 67881 ops/sec;    7.5 MB/s
DB path: [./data]
seekrandom   :     730.041 micros/op 1369 ops/sec;   15.2 MB/s (1000000 of 1000000 found)


real    12m14.817s
user    12m10.258s
sys     0m2.636s
```
**prev-skiplist**:

```
DB path: [./data]
fillseq      :      14.507 micros/op 68930 ops/sec;    7.6 MB/s
DB path: [./data]
seekrandom   :      83.699 micros/op 11947 ops/sec;  132.1 MB/s (1000000 of 1000000 found)


real    1m28.374s
user    1m26.872s
sys     0m1.318s
```