@yiwu-arbug 2019-07-08
Would be great if someone can let us know if we can move forward with this one. 

If it is good to go, we plan to add an option or a separate memtable type to gate the optimization, as the change introduce 8 bytes/key memory overhead to the memtable. We can add a template argument to inlineskiplist (similar to the `UseCAS` argument) to reuse existing code.

For context, we recently find that reverse scan on memtable can be 7-8X slower than forward scan, verse its only 2-3X difference for scan on SST files. We do get users run reverse scan query on recently inserted data, which are likely all in memetable, and they want to improve performance in this scenario.

For correctness of concurrent read/write, some notes:
1. When inserting a key, it is first inserted to reverse linked-list, then to the forward linked-list. We say a key is visible once it is on the forward linked-list.
2. If a reader reach a key by following a `next` link, the key is visible (because of insert order). If a reader reach a key by following a `prev` link, the reader need to double check if the key is visible by double checking if `next` pointer is null.