@siying 2019-07-24
Somehow I don't feel super comfortable with exposing memtable in a inconsistent state. That is, when you go forward, you can see a key, but go backward, it's missing. It's very likely that merging iterator or db_iter is able to handle it, but it feels nicer if the memtable provides a stronger guarantee.