@yiwu-arbug 2019-07-24
@siying appreciate your feedback! Me and @Little-Wallace are actually writing a docs for the patch and here is it. The docs actually patching the issue where range scan can get inconsistent result. See the explanation for `Prev()`. Please check.

https://docs.google.com/document/d/1GRksgBDrDiANFpSOAul4UsV4LfdcTpdr0J5UI-gwiiQ/edit?usp=sharing